feat: LeagueList 컴포넌트화
